requirements.txt
aiosqlite
annotated-types
anyio
click
colorama
fastapi
greenlet
h11
idna
pydantic
pydantic_core
sniffio
SQLAlchemy
starlette
typing-inspection
typing_extensions
uvicorn
asyncpg
python-dotenv